Casting change

+++ neuralNetwork.py
@@ -129,7 +129,7 @@ images = images.numpy()
 # plot the images in the batch, along with predicted and true labels
 fig = plt.figure(figsize=(25, 4))
 for idx in np.arange(batch_size):
-    ax = fig.add_subplot(2, batch_size/2, idx+1, xticks=[], yticks=[])
+    ax = fig.add_subplot(2, int(batch_size/2), idx+1, xticks=[], yticks=[])
     ax.imshow(np.squeeze(images[idx]), cmap='gray')
     ax.set_title("{} ({})".format(str(preds[idx].item()), str(labels[idx].item())),
                  color=("green" if preds[idx]==labels[idx] else "red"))
